Added cart functionality

+++ db/migrate/20231129223227_create_vehicles.rb
@@ -2,7 +2,7 @@ class CreateVehicles < ActiveRecord::Migration[7.0]
   def change
     create_table :vehicles do |t|
       t.string :vehicle_name
-      t.float :vehicle_price
+      t.integer :vehicle_price
       t.string :vin_identification
       t.integer :vehicle_year
       t.references :type, null: false, foreign_key: true
